qwweerty

+++ app/src/main/java/uz/texnopos/installment/data/model/Payment.kt
@@ -2,5 +2,5 @@ package uz.texnopos.installment.data.model
 
 data class Payment(
     val order_id:Int,
-    val amount:Long
+    val amount:Double
 )
